Fix incorrect deprecation warnings
When a property is defined in package.json, the config.get() function
returns 'null' when the property is not set instead of 'undefined'. It
appears 'undefined' is reserved for properties that are actually not
defined for the extension.

As a result, users who never set the deprecated export options would
still see deprecation notifications when the extension processed its
config. Fixing the checks seemed better than removing the property
definitions, since I assume the definitions are still useful for users
who did configure the old properties.